requirements.txt
flask==2.3.3
pandas==2.2.1
plotly==5.19.0
matplotlib==3.8.4
gunicorn==21.2.0
requests==2.31.0
pyserial==3.5
